[poaps] Move types to own file and refactor exports
