+++ packages/poaps/src/types/EmailReservationInput.ts
@@ -1,0 +1,11 @@
+/**
+ * Represents the input fields required to reserve a POAP via email.
+ *
+ * @export
+ * @interface EmailReservationInput
+ */
+export interface EmailReservationInput {
+  mintCode: string;
+  email: string;
+  sendEmail?: boolean;
+}
