+++ packages/poaps/src/types/PoapsSortFields.ts
@@ -1,0 +1,12 @@
+/**
+ * Enum to define available fields for sorting POAPs.
+ *
+ * @export
+ * @enum {string}
+ */
+export enum PoapsSortFields {
+  /** Represents sorting by the date when a POAP was minted. */
+  MintedOn = 'minted_on',
+  /** Represents sorting by the ID of a POAP. */
+  Id = 'id',
+}
