+++ packages/poaps/src/types/WalletMintInput.ts
@@ -1,0 +1,10 @@
+/**
+ * Represents the input fields required to mint a POAP for a wallet.
+ *
+ * @export
+ * @interface WalletMintInput
+ */
+export interface WalletMintInput {
+  mintCode: string;
+  address: string;
+}
